fix styled components flicker
React works in a next way: it render something on the server and than
send to browser. on this step react starts loading it's components and
components styles. In case of this you can see small flicker before
component sytles load. The solution is described on the styled
components github page